requirements.txt
asgiref==3.3.1
certifi==2020.11.8
chardet==3.0.4
Django==3.1.4
django-bootstrap-modal-forms==2.0.1
django-crispy-forms==1.10.0
django-filter==2.4.0
django-widget-tweaks==1.4.8
djangorestframework==3.12.2
idna==2.10
importlib-metadata==3.1.0
Markdown==3.3.3
numpy==1.19.4
pandas==1.1.4
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
six==1.15.0
sqlparse==0.4.1
urllib3==1.26.2
zipp==3.4.0
